task_postgres/requirements.txt
pandas==1.1.2
psycopg2-binary==2.8.5